separate iavlviewer command into subcommands
